refactor(dom): move PNG decode and data: URL parsing from rts-dom-bridge to rts-dom
O rasterizador headless da regua de pintura (examples/claude-raster.rs,
examples/claude-paint-dump.rs) faz layout de <img> SEM a ponte -- nenhum dos
dois tem Engine/Registry -- e por isso nao tinha como chamar o descodificador
de PNG do lote V-img, que so existia em rts-dom-bridge::imagem. Uma
implementacao so: o decodificador (src/imagem/png.rs) e o parse de data: URL
(bytes_da_data_url) mudam-se para rts_dom::imagem, com os seus testes; a
ponte (rts-dom-bridge::imagem) fica fina e chama a mesma funcao.

flate2 passa a dependencia REAL do rts-dom (era so de rts-dom-bridge) --
mesmo pin (1.1, rust_backend). O modulo novo continua sem I/O por desenho
(nenhum std::fs::read): o Cargo.toml do rts-dom documenta 'wasm-safe (sem
I/O)' e um std::fs::read aqui entraria no build normal do crate, que os dois
exemplos tambem linkam -- por isso o loader de disco fica nos EXEMPLOS
(proximo commit), nao neste modulo.

Rejeitado: manter o decodificador so na ponte e fazer os exemplos chamarem
rts-dom-bridge -- os exemplos so linkam rts-dom (sem Engine/Registry), e
trazer rts-dom-bridge para eles inverteria a dependencia (o crate de UI
passaria a depender da ponte).

cargo test -p rts-dom-bridge: 7 -> 4 passed, 0 failed (os 3 testes do
decodificador mudam de crate com o codigo).
cargo test -p rts-dom --lib imagem::: 4 novos, 0 falhas.
